Move MatchingConditions/MatchingRoutine to API
