+++ scijava/scijava-ops-api/src/main/java/org/scijava/ops/api/features/MatchingConditions.java
@@ -1,5 +1,5 @@
 
-package org.scijava.ops.engine.impl;
+package org.scijava.ops.api.features;
 
 import java.util.Objects;
