+++ scijava/scijava-ops-api/src/main/java/org/scijava/ops/api/features/MatchingRoutine.java
@@ -1,0 +1,8 @@
+package org.scijava.ops.api.features;
+
+
+public interface MatchingRoutine {
+
+	MatchingResult match(MatchingConditions conditions);
+
+}
